requirements/mmra_requirement.txt
absl-py==2.2.2
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
angle_emb==0.5.6
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
bitsandbytes==0.46.0
boltons==25.0.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
fsspec==2024.6.1
datasets==3.6.0
decorator==4.4.2
deprecation==2.1.0
dill==0.3.8
docker-pycreds==0.4.0
einops==0.8.1
et_xmlfile==2.0.0
eyeD3==0.9.8
filelock==3.13.1
filetype==1.2.0
fonttools==4.58.1
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
hdbscan==0.8.40
hf-xet==1.1.2
hmmlearn==0.3.3
holidays==0.71
huggingface-hub==0.32.3
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
imbalanced-learn==0.13.0
imblearn==0.0
Jinja2==3.1.4
joblib==1.5.1
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==2.1.5
matplotlib==3.10.1
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
multiprocess==0.70.16
narwhals==1.41.0
networkx==3.3
numba==0.61.2
numpy==1.26.0
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pandas==1.5.0
peft==0.15.2
pillow==11.0.0
platformdirs==4.3.8
plotly==6.0.1
pooch==1.8.2
prettytable==3.16.0
proglog==0.1.12
propcache==0.3.1
protobuf==6.31.1
psutil==7.0.0
pyarrow==20.0.0
pyAudioAnalysis==0.3.14
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydub==0.25.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==4.1.0
sentry-sdk==2.29.1
setproctitle==1.3.6
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
soundfile==0.13.1
soxr==0.5.0.post1
sympy==1.13.3
threadpoolctl==3.6.0
tokenizers==0.21.1
torchinfo==1.8.0
tqdm==4.67.1
transformers==4.48.0
triton==2.2.0
typing-inspection==0.4.1
typing_extensions==4.12.2
urllib3==2.4.0
wandb==0.19.11
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.20.0
